+++ common/Files/File.ts
@@ -51,6 +51,7 @@ export interface Workspace {
     scale: number | null;
     gridScale: number | null;
     defaultHeight: number | null;
+    color: string;
 }
 
 /**
@@ -100,6 +101,7 @@ export interface PartialFile {
     },
     scale?: number;
     gridScale?: number;
+    color?: string;
 }
 
 /**
@@ -136,3 +138,8 @@ export const DEFAULT_WORKSPACE_HEIGHT_INCREMENT = 0.1;
  * The minimum height that hexes in a workspace can be.
  */
 export const DEFAULT_WORKSPACE_MIN_HEIGHT = 0.1;
+
+/**
+ * The default color for workspaces.
+ */
+export const DEFAULT_WORKSPACE_COLOR = "#999999";